ci: not restore cache from same PR (#969)
## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

Previously, we restored the cache from the previous task of the current
PR. However, GitHub runs actions not directly on the PR ref but on a ref
merged with the `main` branch. As a result, using the cache from the
current PR often leads to cache misses when there are changes in the
`main` branch.

We would also save some time on saving the PR cache.

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).