fix(rspeedy/core): remove the default `output.iife: false` (#949)
## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

The `output.iife: false` setting has been removed in this patch to
address the "SyntaxError: invalid redefinition of parameter name" error.

This issue occurs in production when a `const` or `let` variable is
declared with the same name as a variable provided by the Lynx Engine
(e.g., `fetch`, `requestAnimationFrame`).

Originally, this setting was introduced to mitigate performance issues
in PrimJS caused by closures. However, since that problem was resolved
years ago, it is now safe to remove this setting.

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [ ] Documentation updated (or not required).