refactor: save dataset on an attribute (#981)
On lynx, the `data-*` attributes have different behaviors than the
HTMLElement has.

The dataset will be treated as properties, the key will not be applied
the camel-case <-> hyphenate name transformation.

Before this commit we use it as a runtime data, but after this commit we
will use encodeURI(JSON.stringify(dataset)) to encode it as a string.

Implement SSR for Web #52