fix(react): memory leak because of cycle reference (#1071)
Different from #1054, this fix avoids disconnecting the
`SnapshotInstance` that is about to be reused, and avoids incompleteness
when it is reused in the `<list/>`.

<!--
  Thank you for submitting a pull request!

We appreciate the time and effort you have invested in making these
changes. Please ensure that you provide enough information to allow
others to review your pull request.

Upon submission, your pull request will be automatically assigned with
reviewers.

If you want to learn more about contributing to this project, please
visit:
https://github.com/lynx-family/lynx-stack/blob/main/CONTRIBUTING.md.
-->

## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [ ] Documentation updated (or not required).

---------

Signed-off-by: Zhiyuan Hong <28915578+hzy@users.noreply.github.com>
Co-authored-by: Yiming Li <yimingli.cs@gmail.com>