feat(testing-library): replace waitSchedule with act (#1182)
<!--
  Thank you for submitting a pull request!

We appreciate the time and effort you have invested in making these
changes. Please ensure that you provide enough information to allow
others to review your pull request.

Upon submission, your pull request will be automatically assigned with
reviewers.

If you want to learn more about contributing to this project, please
visit:
https://github.com/lynx-family/lynx-stack/blob/main/CONTRIBUTING.md.
-->

## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

`act` is an API both available in `react-dom` and `preact`:

- `react-dom`: https://legacy.reactjs.org/docs/test-utils.html#act,
https://legacy.reactjs.org/docs/testing-recipes.html#act
- `preact`:
https://github.com/preactjs/preact/blob/0a3c7b499dad19344eefa2805e5bc0808ab45056/test-utils/src/index.js#L18-L27

It will ensure all state and effect updates are finished before your
test assertion. Before we export a `waitSchedule` function with the same
purpose, but `act` should be the official way to do so.

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [ ] Documentation updated (or not required).