fix(react/runtime): make `loadLazyBundle` return sync `then` (#1212)
## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

Make lazy bundle being able to resolve and rendered on the first screen
of background thread.

The `QueryComponent` API can support both blocking and non-blocking
returns.

**Before this patch**:

- main-thread: Directly render the children of `<Suspense />`, which
resolves to the content of the lazy bundle.
- background: Firstly render to the children of `<Suspense />`, which
throw an error. Then the promise of `loadLazyBundle` resolves, which
makes `<Suspense />` updates to render the content of the lazy bundle.

The render result would be:

Full content -> blank -> full content

which is quite weird for end-user.

**After this patch**:

Both main-thread and background would directly render the children of
`<Suspense />`, which resolves to the content of the lazy bundle.

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [ ] Documentation updated (or **not required**).